Added (int) typecast to SAFECOPY macro to make gcc -Wall happy.

+++ src/xpdev/gen_defs.h
@@ -132,7 +132,7 @@ enum {
 #define HANDLE	void*
 #endif
 
-#define SAFECOPY(dst,src)	sprintf(dst,"%.*s",sizeof(dst)-1,src)
+#define SAFECOPY(dst,src)	sprintf(dst,"%.*s",(int)sizeof(dst)-1,src)
 
 /****************************************************************************/
 /* MALLOC/FREE Macros for various compilers and environments				*/
